updated fastphot data model
