+++ bin/fastspecfit
@@ -1,7 +1,8 @@
 #!/usr/bin/env python
 """fastspecfit for DESI.
 
-Use --mpi for the MPI version.
+fastspecfit /global/cfs/cdirs/desi/spectro/redux/daily/tiles/80608/20201215/zbest-9-80608-20201215.fits -o specfit.fits --ntargets 2 --mp 1
+fastspecfit /global/cfs/cdirs/desi/spectro/redux/daily/tiles/80605/20201215/zbest-9-80605-20201215.fits -o photfit.fits --ntargets 2 --mp 1 --photfit
 
 """
 import os
