+++ bin/mpi-fastspecfit
@@ -81,7 +81,7 @@ def spectra2outfiles(specfiles, inprefix, outprefix, ext=None, outdir=None):
 
     return np.array(outfiles)
 
-def find_specfiles(reduxdir, outdir=None, prefix='zbest', outprefix='specfit',
+def find_specfiles(reduxdir, outdir=None, prefix='zbest', outprefix='fastspec',
                    avoiddir=None):
     '''
     Returns list of spectra files under reduxdir that need to be processed.
@@ -279,10 +279,10 @@ def plan(args, comm=None, merge=False):
             explist = explist[np.isin(explist['TILEID'].astype(str), tiles)]
             #explist = explist[:10]
 
-        if args.photfit:
-            outprefix = 'photfit'
+        if args.fastphot:
+            outprefix = 'fastphot'
         else:
-            outprefix = 'specfit'
+            outprefix = 'fastspec'
 
         specfiles, outfiles = [], []
         alltiles = explist['TILEID'].astype(str)
@@ -364,7 +364,7 @@ def plan(args, comm=None, merge=False):
         print('#!/bin/bash')
         print('#SBATCH -N {}'.format(numnodes))
         print('#SBATCH -q {}'.format(queue))
-        if args.photfit:
+        if args.fastphot:
             print('#SBATCH -J fastspecfit-phot')
         else:
             print('#SBATCH -J fastspecfit-spec')
@@ -413,10 +413,10 @@ def merge_fastspecfit(args, comm=None):
 
     outdir, _, outfiles, _, _ = plan(args, comm=comm, merge=True)
 
-    if args.photfit:
-        prefix = 'photfit'
+    if args.fastphot:
+        prefix = 'fastphot'
     else:
-        prefix = 'specfit'
+        prefix = 'fastspec'
 
     t0 = time.time()        
     out = Table(np.hstack([fitsio.read(outfile) for outfile in outfiles]))#[:5]]))
@@ -480,8 +480,8 @@ def run_fastspecfit(args, comm=None):
 
         if args.exposures:
             cmd += ' --exposures'
-        if args.photfit:
-            cmd += ' --photfit'
+        if args.fastphot:
+            cmd += ' --fastphot'
         if args.solve_vdisp:
             cmd += ' --solve-vdisp'
 
@@ -542,7 +542,7 @@ def main():
     parser.add_argument('--mp', type=int, default=1, help='Number of multiprocessing processes per MPI rank or node.')
     parser.add_argument('--maxnodes', type=int, default=256, help='maximum number of nodes to use')
 
-    parser.add_argument('--photfit', action='store_true', help='Fit just the broadband photometry.')
+    parser.add_argument('--fastphot', action='store_true', help='Fit just the broadband photometry.')
     parser.add_argument('--solve-vdisp', action='store_true', help='Solve for the velocity disperion.')
 
     parser.add_argument('--merge', action='store_true', help='Merge all individual catalogs into one large file.')
